requirements.txt
faiss-cpu
openai
python-dotenv
numpy
flask
flask_cors
requests
redis
gunicorn
markdownify
tiktoken
PyJWT